Migrate from SeaORM to Prisma (#8)
* werkin

* meh

* major work towards seaorm -> prisma migration

* added prisma commands to scripts

* reworked opds api

* werkin

* did some things

* Migration mostly complete

+++ server/src/fs/media_file.rs
@@ -2,10 +2,8 @@ use rocket::http::ContentType;
 use std::str::FromStr;
 use zip::read::ZipFile;
 
-use crate::types::comic::ComicInfo;
-use crate::types::rocket::ImageResponse;
+use crate::types::{errors::ProcessFileError, http::ImageResponse, models::ComicInfo};
 
-use super::error::ProcessFileError;
 use super::{epub::get_epub_page, rar::get_rar_image, zip::get_zip_image};
 
 // FIXME: this needs to change. I really only need the ComicInfo
@@ -46,7 +44,7 @@ pub fn get_content_type_from_mime(mime: &str) -> ContentType {
     })
 }
 
-pub fn get_page(file: &str, page: usize) -> GetPageResult {
+pub fn get_page(file: &str, page: i32) -> GetPageResult {
     let file_name = file.to_string();
 
     if file_name.ends_with(".cbr") {
